--- src/main/java/umc/server/baeksstreetmapserver/user/dto/PatchUserStatusReq.java
@@ -1,4 +1,0 @@
-package umc.server.baeksstreetmapserver.user.dto;
-
-public class PatchUserStatusReq {
-}
