+++ src/main/java/umc/server/baeksstreetmapserver/user/dto/request/PatchNicknameReq.java
@@ -1,4 +1,4 @@
-package umc.server.baeksstreetmapserver.user.dto;
+package umc.server.baeksstreetmapserver.user.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
