+++ src/main/java/umc/server/baeksstreetmapserver/user/dto/request/PatchUserStatusReq.java
@@ -1,0 +1,4 @@
+package umc.server.baeksstreetmapserver.user.dto.request;
+
+public class PatchUserStatusReq {
+}
